test(dynamic): reorganize test layout by source files
