+++ docs/testing/testing.md
@@ -18,11 +18,11 @@
 
 | Metric | Value |
 | --- | ---: |
-| Total test functions | 9,616 |
-| Unit test functions | 9,369 |
+| Total test functions | 9,626 |
+| Unit test functions | 9,379 |
 | E2E test functions | 247 |
-| cmd test functions | 403 |
-| Test files (internal/) | 406 |
+| cmd test functions | 404 |
+| Test files (internal/) | 407 |
 | Test files (cmd/) | 13 |
 | Test files (test/e2e/suite/) | 109 |
 | Tool sub-packages tested | 165 |
@@ -35,7 +35,7 @@
 
 | Pattern | Count | % |
 | --- | ---: | ---: |
-| `TestFunc_Scenario` (2-part) | 8,621 | 89.7% |
+| `TestFunc_Scenario` (2-part) | 8,631 | 89.7% |
 | `TestFunc` (no underscore) | 705 | 7.3% |
 | `TestFunc_Scenario_Expected` (3+ part) | 290 | 3.0% |
 
@@ -47,10 +47,10 @@
 | --- | ---: | ---: | --- |
 | Core packages | 1,592 | 83 | shared runtime packages such as config, GitLab client, OAuth, resources, prompts, and utilities |
 | Tools orchestration | 233 | 8 | registration, meta-tool dispatch, safe mode, validation, markdown, and routing tests |
-| Tool sub-packages (165) | 7,141 | 315 | domain-specific GitLab tool handlers |
+| Tool sub-packages (165) | 7,150 | 316 | domain-specific GitLab tool handlers |
 | E2E integration | 247 | 109 | build-tagged real GitLab integration suite |
-| cmd packages | 403 | 13 | server entry point and developer command utilities |
-| **Total** | **9,616** | **528** |  |
+| cmd packages | 404 | 13 | server entry point and developer command utilities |
+| **Total** | **9,626** | **529** |  |
 
 ### Core Packages
 
@@ -86,9 +86,9 @@
 | groups | 123 | 98.9% | 18 |
 | jobs | 118 | 96.1% | 17 |
 | search | 114 | 100.0% | 10 |
+| dynamic | 111 | 99.8% | 4 |
 | packages | 108 | 95.6% | 9 |
 | awardemoji | 106 | 65.0% | 25 |
-| dynamic | 102 | 99.8% | 4 |
 | pipelines | 99 | 97.4% | 12 |
 | resourceevents | 99 | 100.0% | 16 |
 | commits | 97 | 96.5% | 13 |
@@ -148,7 +148,7 @@
 | deploytokens | 65 | 2 | 100.0% | 10 |
 | dockerfiletemplates | 14 | 1 | 100.0% | 2 |
 | dorametrics | 9 | 2 | 100.0% | 2 |
-| dynamic | 102 | 6 | 99.8% | 4 |
+| dynamic | 111 | 7 | 99.8% | 4 |
 | elicitationtools | 57 | 2 | 98.2% | 4 |
 | enterpriseusers | 33 | 3 | 100.0% | 4 |
 | environments | 47 | 2 | 100.0% | 6 |
@@ -276,7 +276,7 @@
 | vulnerabilities | 52 | 3 | 98.5% | 8 |
 | wikis | 58 | 2 | 98.9% | 6 |
 | workitems | 66 | 2 | 100.0% | 5 |
-| **Total** | **7,141** | **315** |  | **1,061** |
+| **Total** | **7,150** | **316** |  | **1,061** |
 
 </details>
 
@@ -288,7 +288,7 @@
 | --- | ---: |
 | cmd/audit_godocs | 50.7% |
 | cmd/audit_meta_schema | 80.2% |
-| cmd/audit_metrics | 13.8% |
+| cmd/audit_metrics | 19.7% |
 | cmd/audit_output | 23.0% |
 | cmd/audit_tokens | 19.0% |
 | cmd/eval_meta_tools | 57.3% |
@@ -492,9 +492,9 @@
 Coverage target: **>90%** per package. Packages below the target in the latest generated coverage snapshot:
 
 - **cmd/gen_llms** (7.2%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
-- **cmd/audit_metrics** (13.8%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
 - **cmd/gen_readme** (14.6%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
 - **cmd/audit_tokens** (19.0%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
+- **cmd/audit_metrics** (19.7%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
 - **cmd/gen_testing_docs** (20.7%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
 - **cmd/audit_output** (23.0%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
 - **cmd/audit_godocs** (50.7%) - developer command formatting and reporting branches are covered by focused unit tests plus manual/CI tooling runs.
