+++ internal/tools/dynamic/metrics_test.go
@@ -1,0 +1,23 @@
+package dynamic
+
+import "testing"
+
+func TestMetrics_RuntimeCountersClampNegativeSuppressions(t *testing.T) {
+	ResetSearchRuntimeMetrics()
+	recordSearchRuntimeMetrics(1, false, false, false, -5)
+
+	metrics := SearchRuntimeMetricsSnapshot()
+	if metrics.Searches != 1 {
+		t.Fatalf("Searches = %d, want 1", metrics.Searches)
+	}
+	if metrics.DestructiveFuzzySuppressions != 0 {
+		t.Fatalf("DestructiveFuzzySuppressions = %d, want 0", metrics.DestructiveFuzzySuppressions)
+	}
+}
+
+func TestMetrics_SearchIndexPostingCount(t *testing.T) {
+	index := searchIndex{byToken: map[string][]int{"project": []int{0, 2}, "delete": []int{1}}}
+	if got := searchIndexPostingCount(index); got != 3 {
+		t.Fatalf("searchIndexPostingCount() = %d, want 3", got)
+	}
+}
