+++ internal/tools/groupmilestones/markdown.go
@@ -37,10 +37,10 @@ func FormatMarkdown(v Output) string {
 		fmt.Fprintf(&b, toolutil.FmtMdUpdated, toolutil.FormatTime(v.UpdatedAt))
 	}
 	toolutil.WriteHints(&b,
-		"If the workflow asks to fetch/get before update or delete, use the selected tool surface's group-milestone get action with the same group_id and this milestone_id next",
-		"Use the selected tool surface's group-milestone update action with the same group_id and this milestone_id to modify this milestone",
-		"Use the selected tool surface's group-milestone issues or merge-requests action with the same group_id and milestone_id to list associated items",
-		"Use the selected tool surface's group-milestone delete action with the same group_id, this milestone_id, and explicit confirm=true to remove this milestone",
+		"Use gitlab_group_milestone_get with the same group_id and milestone_id before update/delete workflows",
+		"Use gitlab_group_milestone_update with the same group_id and milestone_id to modify this milestone",
+		"Use gitlab_group_milestone_issues or gitlab_group_milestone_merge_requests with the same group_id and milestone_id to list associated items",
+		"Use gitlab_group_milestone_delete with the same group_id, milestone_id, and confirm=true to remove this milestone",
 	)
 	return b.String()
 }
@@ -62,8 +62,9 @@ func FormatListMarkdownString(out ListOutput) string {
 	}
 	toolutil.WritePagination(&b, out.Pagination)
 	toolutil.WriteHints(&b,
-		"Use the selected tool surface's group-milestone get action with the same group_id and milestone_id for full details before update/delete workflows",
-		"Use the selected tool surface's group-milestone create action with group_id to add a new group milestone",
+		toolutil.HintPreserveLinks,
+		"Use gitlab_group_milestone_get with group_id and milestone_id for full details before update/delete",
+		"Use gitlab_group_milestone_create with group_id to add a new group milestone",
 	)
 	return b.String()
 }
